Move to XCode 4.6
Move to XCode 4.6 and some code clean up. More object interfaces to
follow. Also ObjC can delivery versus the CCArray interface.

+++ Sim/CCArray.m
@@ -114,7 +114,7 @@
 {
 	if( index >= data->num )
 		[NSException raise:NSRangeException
-					format: @"index out of range in objectAtIndex(%d)", data->num ];
+					format: @"index out of range in objectAtIndex(%ld)", (unsigned long)data->num ];
 	
 	return data->arr[index];
 }
